+++ bs/display/widget/bsText.h
@@ -196,7 +196,7 @@ static NSDictionary* __bsText_keyValues = nil;
     NSMutableArray *remain = [NSMutableArray array];
     NSInteger f0 = 0;
     CGFloat fontSize = 0;
-    for( int i = 0, j = [params count]; i < j; ) {
+    for( NSInteger i = 0, j = [params count]; i < j; ) {
         NSString *k = (NSString*)params[i++];
         NSString *v = (NSString*)params[i++];
         NSInteger num = [[__bsText_keyValues objectForKey:k] integerValue];
